# Conflicts:
#	.idea/deploymentTargetSelector.xml
#	app/src/androidTest/java/es/uji/smallaris/model/TestServicioRutas.kt
#	app/src/androidTest/java/es/uji/smallaris/model/TestServicioVehiculos.kt
#	app/src/main/java/es/uji/smallaris/model/CalculadorRutas.kt
#	app/src/main/java/es/uji/smallaris/model/CalculadorRutasORS.kt
#	app/src/main/java/es/uji/smallaris/model/RepositorioFirebase.kt
#	app/src/main/java/es/uji/smallaris/model/RutaBuilder.kt
#	app/src/main/java/es/uji/smallaris/model/ServicioAPIs.kt
#	app/src/main/java/es/uji/smallaris/model/ServicioRutas.kt

+++ app/src/main/java/es/uji/smallaris/model/Ruta.kt
@@ -10,7 +10,7 @@ class Ruta(
     private val trayecto: LineString,
     private val distancia: Float,
     private val duracion: Float,
-    private val coste: Float
+    private val coste: Double
 ) {
     fun getTrayecto(): LineString {
         return trayecto
@@ -24,7 +24,7 @@ class Ruta(
         return duracion
     }
 
-    fun getCoste(): Float {
+    fun getCoste(): Double {
         return coste
     }
 }
